Moved WorkflowOperationException and its children to cadence.workflow
