+++ src/main/java/com/uber/cadence/internal/sync/ActivityInvocationHandler.java
@@ -19,7 +19,7 @@ package com.uber.cadence.internal.sync;
 import com.google.common.base.Defaults;
 import com.uber.cadence.activity.ActivityMethod;
 import com.uber.cadence.activity.MethodRetry;
-import com.uber.cadence.internal.ActivityException;
+import com.uber.cadence.workflow.ActivityException;
 import com.uber.cadence.internal.common.InternalUtils;
 import com.uber.cadence.activity.ActivityOptions;
 import com.uber.cadence.workflow.Promise;
