+++ src/main/java/com/uber/cadence/internal/sync/SyncDecisionContext.java
@@ -22,8 +22,8 @@ import com.uber.cadence.WorkflowType;
 import com.uber.cadence.activity.ActivityOptions;
 import com.uber.cadence.common.RetryOptions;
 import com.uber.cadence.converter.DataConverter;
-import com.uber.cadence.internal.ActivityException;
-import com.uber.cadence.internal.ChildWorkflowTaskFailedException;
+import com.uber.cadence.workflow.ActivityException;
+import com.uber.cadence.internal.replay.ChildWorkflowTaskFailedException;
 import com.uber.cadence.internal.WorkflowRetryerInternal;
 import com.uber.cadence.internal.replay.ActivityTaskFailedException;
 import com.uber.cadence.internal.replay.ExecuteActivityParameters;
