+++ src/main/java/com/uber/cadence/workflow/SignalExternalWorkflowException.java
@@ -14,9 +14,10 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal;
+package com.uber.cadence.workflow;
 
 import com.uber.cadence.WorkflowExecution;
+import com.uber.cadence.workflow.WorkflowOperationException;
 
 /**
  * Exception used to communicate failure of a request to signal an external workflow.
