+++ src/main/java/com/uber/cadence/workflow/StartChildWorkflowFailedException.java
@@ -14,7 +14,7 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal;
+package com.uber.cadence.workflow;
 
 import com.uber.cadence.ChildWorkflowExecutionFailedCause;
 import com.uber.cadence.WorkflowExecution;
