+++ src/main/java/com/uber/cadence/workflow/WorkflowOperationException.java
@@ -14,7 +14,7 @@
  *  express or implied. See the License for the specific language governing
  *  permissions and limitations under the License.
  */
-package com.uber.cadence.internal;
+package com.uber.cadence.workflow;
 
 /**
  * Base exception used to communicate a failure that can be thrown by operations requested by a workflow code.
